upd patch and del /conts routes

+++ src/services/contacts.js
@@ -53,13 +53,13 @@ export const createContact = async (payload) => {
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
-    const rawResult = await ContactsCollection.findOneAndUpdate({ _id: contactId }, payload, { new: true, includeResultMetadata: true, ...options });
+export const updateContact = async (userId, contactId, payload, options = {}) => {
+    const rawResult = await ContactsCollection.findOneAndUpdate({ _id: contactId, userId }, payload, { new: true, includeResultMetadata: true, ...options });
     if (!rawResult || !rawResult.value) return null;
     return rawResult.value;
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (userId, contactId) => {
+    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId, userId });
     return contact;
 };
